Add prompt when deleting file, fixed a bug with the keycloak token prompt always overriding the body text and added option to show different type of buttons in the prompt

+++ projects/valtimo/dossier/src/lib/dossier-detail/tab/documenten-api-documents/documenten-api-documents.component.ts
@@ -25,6 +25,7 @@ import {TranslateService} from '@ngx-translate/core';
 import {ConfigService} from '@valtimo/config';
 import {DocumentenApiMetadata} from '@valtimo/components';
 import {UserProviderService} from '@valtimo/security';
+import {PromptService} from '@valtimo/user-interface';
 
 @Component({
   selector: 'valtimo-dossier-detail-tab-documenten-api-documents',
@@ -87,6 +88,7 @@ export class DossierDetailTabDocumentenApiDocumentsComponent implements OnInit {
     private readonly toastrService: ToastrService,
     private readonly uploadProviderService: UploadProviderService,
     private readonly downloadService: DownloadService,
+    private readonly promptService: PromptService,
     private readonly translateService: TranslateService,
     private readonly configService: ConfigService,
     private readonly userProviderService: UserProviderService
@@ -136,16 +138,30 @@ export class DossierDetailTabDocumentenApiDocumentsComponent implements OnInit {
       });
   }
 
-  removeRelatedFile(relatedFile: RelatedFile): void {
-    this.documentService.removeResource(this.documentId, relatedFile.fileId).subscribe(
-      () => {
-        this.toastrService.success('Successfully removed document from dossier');
-        this.refetchDocuments();
-      },
-      () => {
-        this.toastrService.error('Failed to remove document from dossier');
+  removeRelatedFile(relatedFile: RelatedFile) {
+    this.promptService.openPrompt({
+      headerText: this.translateService.instant('dossier.deleteConfirmation.title'),
+      bodyText: this.translateService.instant('dossier.deleteConfirmation.description'),
+      cancelButtonText: this.translateService.instant('dossier.deleteConfirmation.cancel'),
+      confirmButtonText: this.translateService.instant('dossier.deleteConfirmation.delete'),
+      cancelMdiIcon: 'cancel',
+      confirmMdiIcon: 'delete',
+      cancelButtonType: 'secondary',
+      confirmButtonType: 'primary',
+      closeOnConfirm: true,
+      closeOnCancel: true,
+      confirmCallBackFunction: () => {
+        this.documentService.removeResource(this.documentId, relatedFile.fileId).subscribe(
+          () => {
+            this.toastrService.success('Successfully removed document from dossier');
+            this.refetchDocuments();
+          },
+          () => {
+            this.toastrService.error('Failed to remove document from dossier');
+          }
+        );
       }
-    );
+    })
   }
 
   metadataSet(metadata: DocumentenApiMetadata): void {
